HtmlConverter

missing description

Version: 0.1.0

Generated: 2022-05-24T11:50:20.685

Arguments

fontfiles
fontfiles - a character vector of font files that will be imported into the converter. - (defaulting to extrafont::fonttable()$fontfile) - (java expects a RCharacterVector)

Details

no details

Methods

Constructors

J$HtmlConverter$new(fontfiles)

Static methods

none

Instance methods

instance$urlToPdf(htmlUrl, outFile)

instance$urlComponentToPdf(htmlUrl, outFile, cssSelector)

instance$fileToPdf(inFile, outFile)

instance$stringToPdf(html, outFile, baseUri)

instance$fitIntoPage(htmlFragment, outFile, maxWidthInches, maxHeightInches, formats, pngDpi)

instance$fitIntoA4(htmlFragment, outFile, xMarginInInches, yMarginInInches, formats)

instance$clone()

instance$print()

Method new()

Create a new HtmlConverter for creating PDF and PNG files from HTML.

Usage

J = html2pdfr::JavaApi$get()
J$HtmlConverter$new(fontfiles)
        

Arguments

fontfiles - a character vector of font files that will be imported into the converter. - (defaulting to extrafont::fonttable()$fontfile) - (java expects a RCharacterVector)

Returns

R6 HtmlConverter object:

Examples

not available

Method urlToPdf()

Convert HTML from a URL to a PDF file. PDF size will be controlled by page media directives within the html.

Usage

J = html2pdfr::JavaApi$get()
instance = J$HtmlConverter$new(...);
instance$urlToPdf(htmlUrl, outFile)
        

Arguments

htmlUrl the URL - (java expects a String)

outFile the full path of the output file - (java expects a String)

Returns

RCharacter: the filename written to (with extension '.pdf' if outFile did not have an extension).

Examples

not available

Method urlComponentToPdf()

Convert HTML from a URL to a PDF file. PDF size will be controlled by page media directives within the html.

Usage

J = html2pdfr::JavaApi$get()
instance = J$HtmlConverter$new(...);
instance$urlComponentToPdf(htmlUrl, outFile, cssSelector)
        

Arguments

htmlUrl the URL - (java expects a String)

outFile the full path of the output file - (java expects a String)

cssSelector the part of the page you want to convert to PDF. - (java expects a String)

Returns

RCharacter: the filename written to (with extension '.pdf' if outFile did not have an extension).

Examples

not available

Method fileToPdf()

Convert HTML from a local file to a PDF file. PDF size will be controlled by page media directives within the html.

Usage

J = html2pdfr::JavaApi$get()
instance = J$HtmlConverter$new(...);
instance$fileToPdf(inFile, outFile)
        

Arguments

inFile the full path to an input HTML file - (java expects a String)

outFile the full path to the output pdf file. (N.B. this function can also output PNG if specified in the filename extension) - (java expects a String)

Returns

RCharacter: the filename written to (with extension '.pdf' if outFile did not have an extension).

Examples

not available

Method stringToPdf()

no description

Usage

J = html2pdfr::JavaApi$get()
instance = J$HtmlConverter$new(...);
instance$stringToPdf(html, outFile, baseUri)
        

Arguments

html the HTML string - (java expects a String)

outFile the full path to the output pdf file (N.B. this function can also output PNG if specified in the filename extension) - (java expects a String)

baseUri optionally the base URI of the HTML string for resolving relative URLs in the HTML (e.g. CSS files). - (defaulting to NA_character_) - (java expects a RCharacter)

Returns

RCharacter: the filename written to (with extension '.pdf' if outFile did not have an extension).

Examples

not available

Method fitIntoPage()

Render HTML string to fit into a page,

Usage

J = html2pdfr::JavaApi$get()
instance = J$HtmlConverter$new(...);
instance$fitIntoPage(htmlFragment, outFile, maxWidthInches, maxHeightInches, formats, pngDpi)
        

Arguments

htmlFragment a HTML fragment, e.g. the table element. It is usually expected there will not be any page media directives in the HTML - (java expects a String)

outFile the full path to the output pdf file (N.B. this function can also output PNG if specified in the filename extension) - (java expects a String)

maxWidthInches what is the maximum allowable width? - (defaulting to 6.25) - (java expects a double)

maxHeightInches what is the maximium allowable height? (if the content is larger than this then it will overflow to another page) - (defaulting to 9.75) - (java expects a double)

formats If the outFile does not specify a file extension then you can do so here as "png" or "pdf" or both. - (defaulting to c('pdf','png')) - (java expects a RCharacterVector)

pngDpi if outputting a PNG the dpi will determine the dimensions of the image. - (defaulting to 300) - (java expects a double)

Returns

RCharacterVector: the filename(s) written to (with extension '.pdf' and '.png' if outFile did not have an extension).

Examples

not available

Method fitIntoA4()

Render HTML string to fit into an A4 page,

Usage

J = html2pdfr::JavaApi$get()
instance = J$HtmlConverter$new(...);
instance$fitIntoA4(htmlFragment, outFile, xMarginInInches, yMarginInInches, formats)
        

Arguments

htmlFragment a HTML fragment, e.g. the table element. It is usually expected there will not be any page media directives in the HTML - (java expects a String)

outFile the full path to the output pdf file (N.B. this function can also output PNG if specified in the filename extension) - (java expects a String)

xMarginInInches page margins - (defaulting to 1.0) - (java expects a double)

yMarginInInches page margins - (defaulting to 1.0) - (java expects a double)

formats If the outFile does not specify a file extension then you can do so here as "png" or "pdf" or both. - (defaulting to c('pdf','png')) - (java expects a RCharacterVector)

Returns

RCharacterVector: the filename(s) written to (with extension '.pdf' and '.png' if outFile did not have an extension).

Examples

not available

Examples

if (FALSE) {
J = html2pdfr::JavaApi$get();
instance = J$HtmlConverter$new(fontfiles)

## -----------------------------------
## Method `HtmlConverter$urlToPdf`
## -----------------------------------

## -----------------------------------
## Method `HtmlConverter$urlComponentToPdf`
## -----------------------------------

## -----------------------------------
## Method `HtmlConverter$fileToPdf`
## -----------------------------------

## -----------------------------------
## Method `HtmlConverter$stringToPdf`
## -----------------------------------

## -----------------------------------
## Method `HtmlConverter$fitIntoPage`
## -----------------------------------

## -----------------------------------
## Method `HtmlConverter$fitIntoA4`
## -----------------------------------

}