Convert HTML document from a local file to a PDF document.

The HTML in `inFile` is assumed to be a complete document. Relative references are resolved with reference to the HTML file on the file system, so correctly located images etc whould be picked up without requiring a server. The resulting PDF size will be controlled by page media directives within the HTML, unless explicitly given here in `maxWidthInches` and `maxHeightInches`. If the `cssSelector` parameter is given the HTML fragment at that selector will be used. In this case it is will be resized to fit within the given dimensions and shrink wrapped so that the content is smaller. If no dimensions are present this will default to A4.

file_to_pdf(
  inFile,
  outFile=tempfile('html2pdfr_'),
  cssSelector=NA_character_,
  xMarginInches=NA_real_,
  yMarginInches=NA_real_,
  maxWidthInches=NA_real_,
  maxHeightInches=NA_real_,
  formats=c('pdf'),
  pngDpi=300,
  converter=html2pdfr::html_converter()
)

Arguments

inFile
inFile the full path the the HTML file - (java expects a String)

outFile
outFile the full path of the output file - (defaulting to tempfile('html2pdfr_')) - (java expects a String)

cssSelector
cssSelector the part of the page you want to convert to PDF. - (defaulting to NA_character_) - (java expects a RCharacter)

xMarginInches
xMarginInches page width margins - (defaulting to NA_real_) - (java expects a RNumeric)

yMarginInches
yMarginInches page height margins - (defaulting to NA_real_) - (java expects a RNumeric)

maxWidthInches
maxWidthInches what is the maximum allowable width? - (defaulting to NA_real_) - (java expects a RNumeric)

maxHeightInches
maxHeightInches what is the maximum allowable height? (if the content is larger than this then it will overflow to another page) - (defaulting to NA_real_) - (java expects a RNumeric)

formats
formats If the outFile does not specify a file extension then you can do so here as "png" or "pdf" or both. - (defaulting to c('pdf')) - (java expects a RCharacterVector)

pngDpi
pngDpi the dots per inch for png outputs if requested - (defaulting to 300) - (java expects a RNumeric)

converter
converter (optional) a configured HTML converter, only needed if manually specifying fonts. - (defaulting to html2pdfr::html_converter()) - (java expects a HtmlConverter)

Value

RCharacterVector: the filename written to (with extension '.pdf' or '.png' if outFile did not have an extension).

Examples

library(testthat)
dest = tempfile(fileext='.html')
download.file('https://cran.r-project.org/banner.shtml', destfile = dest)
file_to_pdf(dest)

#> [1] "/tmp/RtmpWuxDS3/html2pdfr_29b557bbf517.pdf"