HtmlConverter

missing description

Version: 0.4.0

Generated: 2022-07-07T16:11:14.493

Arguments

fontfiles

fontfiles - (java expects a String)

Details

no details

Methods

Constructors

J$HtmlConverter$new(fontfiles)

Static methods

J$HtmlConverter$htmlConverter(fontfiles, update)

J$HtmlConverter$urlToPdf(htmlUrl, outFile, cssSelector, xMarginInches, yMarginInches, maxWidthInches, maxHeightInches, formats, pngDpi, converter)

J$HtmlConverter$fileToPdf(inFile, outFile, cssSelector, xMarginInches, yMarginInches, maxWidthInches, maxHeightInches, formats, pngDpi, converter)

J$HtmlConverter$htmlDocumentToPdf(html, outFile, baseUri, cssSelector, xMarginInches, yMarginInches, maxWidthInches, maxHeightInches, formats, pngDpi, converter)

J$HtmlConverter$htmlFragmentToPdf(htmlFragment, outFile, xMarginInches, yMarginInches, maxWidthInches, maxHeightInches, formats, pngDpi, converter)

Instance methods

none

instance$clone()

instance$print()

Method new()

the default no-args constructor

Usage

J = html2pdfr::JavaApi$get()
J$HtmlConverter$new(fontfiles)
        

Arguments

fontfiles - (java expects a String)

Returns

R6 HtmlConverter object:

Examples

not available

Method htmlConverter()

Create a new HtmlConverter

for creating PDF and PNG files from HTML. In general this will be created automatically. but if you have specific fonts you want to use then you may need to pass them to this function and specify the result in the `converter` parameter of the main functions.

Usage

J = html2pdfr::JavaApi$get()
J$HtmlConverter$htmlConverter(fontfiles, update)
        

Arguments

fontfiles - a character vector of font files that will be imported into the converter. - (defaulting to systemfonts::system_fonts()$path) - (java expects a RCharacterVector)

update - (defaulting to FALSE) - (java expects a RLogical)

Returns

R6 HtmlConverter object:

Examples

 
      conv = html2pdfr::html_converter()
      

Method urlToPdf()

Convert HTML document from a URL to a PDF document.

The URL is assumed to be a complete document. The resulting PDF size will be controlled by page media directives within the HTML, unless explicitly given here in `maxWidthInches` and `maxHeightInches`. If the `cssSelector` parameter is given the HTML fragment at that selector will be used. In this case it is will be resized to fit within the given dimensions and shrink wrapped so that the content is smaller. If no dimensions are present this will default to A4.

Usage

J = html2pdfr::JavaApi$get()
J$HtmlConverter$urlToPdf(htmlUrl, outFile, cssSelector, xMarginInches, yMarginInches, maxWidthInches, maxHeightInches, formats, pngDpi, converter)
        

Arguments

htmlUrl the URL - (java expects a String)

outFile the full path of the output file - (defaulting to tempfile('html2pdfr_')) - (java expects a String)

cssSelector the part of the page you want to convert to PDF. - (defaulting to NA_character_) - (java expects a RCharacter)

xMarginInches page width margins - (defaulting to NA_real_) - (java expects a RNumeric)

yMarginInches page height margins - (defaulting to NA_real_) - (java expects a RNumeric)

maxWidthInches what is the maximum allowable width? - (defaulting to NA_real_) - (java expects a RNumeric)

maxHeightInches what is the maximum allowable height? (if the content is larger than this then it will overflow to another page) - (defaulting to NA_real_) - (java expects a RNumeric)

formats If the outFile does not specify a file extension then you can do so here as "png" or "pdf" or both. - (defaulting to c('pdf')) - (java expects a RCharacterVector)

pngDpi the dots per inch for png outputs if requested - (defaulting to 300) - (java expects a RNumeric)

converter (optional) a configured HTML converter, only needed if manually specifying fonts. - (defaulting to html2pdfr::html_converter()) - (java expects a HtmlConverter)

Returns

RCharacterVector: the filename written to (with extension '.pdf' or '.png' if outFile did not have an extension).

Examples

url_to_pdf('https://cran.r-project.org/banner.shtml')
      

Method fileToPdf()

Convert HTML document from a local file to a PDF document.

The HTML in `inFile` is assumed to be a complete document. Relative references are resolved with reference to the HTML file on the file system, so correctly located images etc whould be picked up without requiring a server. The resulting PDF size will be controlled by page media directives within the HTML, unless explicitly given here in `maxWidthInches` and `maxHeightInches`. If the `cssSelector` parameter is given the HTML fragment at that selector will be used. In this case it is will be resized to fit within the given dimensions and shrink wrapped so that the content is smaller. If no dimensions are present this will default to A4.

Usage

J = html2pdfr::JavaApi$get()
J$HtmlConverter$fileToPdf(inFile, outFile, cssSelector, xMarginInches, yMarginInches, maxWidthInches, maxHeightInches, formats, pngDpi, converter)
        

Arguments

inFile the full path the the HTML file - (java expects a String)

outFile the full path of the output file - (defaulting to tempfile('html2pdfr_')) - (java expects a String)

cssSelector the part of the page you want to convert to PDF. - (defaulting to NA_character_) - (java expects a RCharacter)

xMarginInches page width margins - (defaulting to NA_real_) - (java expects a RNumeric)

yMarginInches page height margins - (defaulting to NA_real_) - (java expects a RNumeric)

maxWidthInches what is the maximum allowable width? - (defaulting to NA_real_) - (java expects a RNumeric)

maxHeightInches what is the maximum allowable height? (if the content is larger than this then it will overflow to another page) - (defaulting to NA_real_) - (java expects a RNumeric)

formats If the outFile does not specify a file extension then you can do so here as "png" or "pdf" or both. - (defaulting to c('pdf')) - (java expects a RCharacterVector)

pngDpi the dots per inch for png outputs if requested - (defaulting to 300) - (java expects a RNumeric)

converter (optional) a configured HTML converter, only needed if manually specifying fonts. - (defaulting to html2pdfr::html_converter()) - (java expects a HtmlConverter)

Returns

RCharacterVector: the filename written to (with extension '.pdf' or '.png' if outFile did not have an extension).

Examples

dest = tempfile(fileext='.html')
download.file('https://cran.r-project.org/banner.shtml', destfile = dest)
file_to_pdf(dest)
      

Method htmlDocumentToPdf()

Convert HTML document from a string to a PDF document.

The HTML in `html` is assumed to be a complete document. Relative references are resolved with reference to `baseUri` if it is given (which could be a `file://` URI). The resulting PDF size will be controlled by page media directives within the HTML, unless explicitly given here in `maxWidthInches` and `maxHeightInches`. If the `cssSelector` parameter is given the HTML fragment at that selector will be used. In this case it is will be resized to fit within the given dimensions and shrink wrapped so that the content is smaller. If no dimensions are present this will default to A4.

Usage

J = html2pdfr::JavaApi$get()
J$HtmlConverter$htmlDocumentToPdf(html, outFile, baseUri, cssSelector, xMarginInches, yMarginInches, maxWidthInches, maxHeightInches, formats, pngDpi, converter)
        

Arguments

html the html document as a string - (java expects a String)

outFile the full path of the output file - (defaulting to tempfile('html2pdfr_')) - (java expects a String)

baseUri the URI from which to interpret relative links in the html content. - (defaulting to NA_character_) - (java expects a RCharacter)

cssSelector the part of the page you want to convert to PDF. - (defaulting to NA_character_) - (java expects a RCharacter)

xMarginInches page width margins - (defaulting to NA_real_) - (java expects a RNumeric)

yMarginInches page height margins - (defaulting to NA_real_) - (java expects a RNumeric)

maxWidthInches what is the maximum allowable width? - (defaulting to NA_real_) - (java expects a RNumeric)

maxHeightInches what is the maximum allowable height? (if the content is larger than this then it will overflow to another page) - (defaulting to NA_real_) - (java expects a RNumeric)

formats If the outFile does not specify a file extension then you can do so here as "png" or "pdf" or both. - (defaulting to c('pdf')) - (java expects a RCharacterVector)

pngDpi the dots per inch for png outputs if requested - (defaulting to 300) - (java expects a RNumeric)

converter (optional) a configured HTML converter, only needed if manually specifying fonts. - (defaulting to html2pdfr::html_converter()) - (java expects a HtmlConverter)

Returns

RCharacterVector: the filename written to (with extension '.pdf' or '.png' if outFile did not have an extension).

Examples

library(tidyverse)
html = readr::read_file('https://fred-wang.github.io/MathFonts/mozilla_mathml_test/')
html_document_to_pdf(html, baseUri = 'https://fred-wang.github.io/MathFonts/mozilla_mathml_test/')
      

Method htmlFragmentToPdf()

Render HTML fragment from a string to a PDF image.

This is the simple rendering function that will output a PDF file (potentially many pages) and a set of PNG files from HTML content. This is primarily used to render HTML content (e.g. a table) that is being included in a larger document. In this case the HTML fragment will not specify page dimensions which need to be provided (defaults to A4 size with 1 inch margins). The result can be embedded into an existing page using latex's includegraphics directive exactly the same way as a graphical figure might be used. The sizing of the output will always be smaller than the dimensions of a page, but will shrink to fit the content.

Usage

J = html2pdfr::JavaApi$get()
J$HtmlConverter$htmlFragmentToPdf(htmlFragment, outFile, xMarginInches, yMarginInches, maxWidthInches, maxHeightInches, formats, pngDpi, converter)
        

Arguments

htmlFragment a HTML fragment, e.g. usually the table element, but may be the whole page. - (java expects a String)

outFile the full path with or without extension (if no extension specified then `formats` parameter will apply) - (defaulting to tempfile('html2pdfr_')) - (java expects a String)

xMarginInches page width margins - (defaulting to 1.0) - (java expects a RNumeric)

yMarginInches page height margins - (defaulting to 1.0) - (java expects a RNumeric)

maxWidthInches what is the maximum allowable width? (default is A4) - (defaulting to 6.25) - (java expects a RNumeric)

maxHeightInches what is the maximum allowable height? (if the content is larger than this then it will overflow to another page) - (defaulting to 9.75) - (java expects a RNumeric)

formats If the outFile does not specify a file extension then you can do so here as "png" or "pdf" or both. - (defaulting to c('pdf','png')) - (java expects a RCharacterVector)

pngDpi the dots per inch for png outputs if requested. - (defaulting to 300) - (java expects a RNumeric)

converter (optional) a configured HTML converter, only needed if manually specifying fonts. - (defaulting to html2pdfr::html_converter()) - (java expects a HtmlConverter)

Returns

RCharacterVector: the filename(s) written to (with extension '.pdf' or '.png' if outFile did not have an extension).

Examples

library(tidyverse)
html = iris 
  huxtable::as_hux() 
html_fragment_to_pdf(html)
      

Examples

if (FALSE) {
J = html2pdfr::JavaApi$get();
instance = J$HtmlConverter$new(fontfiles)

## -----------------------------------
## Method `HtmlConverter$htmlConverter`
## -----------------------------------
 
      conv = html2pdfr::html_converter()

## -----------------------------------
## Method `HtmlConverter$urlToPdf`
## -----------------------------------
url_to_pdf('https://cran.r-project.org/banner.shtml')

## -----------------------------------
## Method `HtmlConverter$fileToPdf`
## -----------------------------------
dest = tempfile(fileext='.html')
download.file('https://cran.r-project.org/banner.shtml', destfile = dest)
file_to_pdf(dest)

## -----------------------------------
## Method `HtmlConverter$htmlDocumentToPdf`
## -----------------------------------
library(tidyverse)
html = readr::read_file('https://fred-wang.github.io/MathFonts/mozilla_mathml_test/')
html_document_to_pdf(html, baseUri = 'https://fred-wang.github.io/MathFonts/mozilla_mathml_test/')

## -----------------------------------
## Method `HtmlConverter$htmlFragmentToPdf`
## -----------------------------------
library(tidyverse)
html = iris 
  huxtable::as_hux() 
html_fragment_to_pdf(html)

}